third_party/WebKit/LayoutTests/inspector/console/console-dir.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/console-test.js"></script>
<script>

function onload()
{
    console.dir(["test1", "test2"]);
    console.dir(document.childNodes);
    console.dir(document.evaluate("//head", document, null, XPathResult.ANY_TYPE, null));

    // Object with properties containing whitespaces
    var obj = { $foo5_: 0 };
    obj[" a b "] = " a b ";
    obj["c d"] = "c d";
    obj[""] = "";
    obj["  "] = "  ";
    obj["a\n\nb\nc"] = "a\n\nb\nc";
    console.dir(obj);

    runTest();
}

function test()
{
    var messages = WebInspector.console.messages;
    for (var i = 0; i < messages.length; ++i) {
        var element = messages[i].toMessageElement();
        var node = element.traverseNextNode(element);
        while (node) {
            if (node._section) {
                messages[i].section = node._section;
                node._section.expanded = true;
                break;
            }
            node = node.traverseNextNode(element);
        }
    }

    InspectorTest.runAfterPendingDispatches(dumpConsoleMessages);

    function dumpConsoleMessages()
    {
        InspectorTest.addResult(InspectorTest.textContentWithLineBreaks(WebInspector.consoleView.topGroup.element));
        InspectorTest.completeTest();
    }
}

</script>
</head>

<body onload="onload()">
<p>
Tests that console logging dumps proper messages.
</p>

</body>
</html>
